:recycle: [Refactor] RepositoryTestSupport 구현 및 리팩토링
